requirements.txt
Flask==3.0.0
paho-mqtt==1.6.1
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
